fix: 收紧 RSS 工具上下文边界
