+++ qq-maid-core/src/app/runtime.rs
@@ -23,8 +23,8 @@ use crate::{
         knowledge::KnowledgeIndex,
         memory::MemoryStore,
         prompt::PromptConfig,
-        rss::{RssFetchConfig, RssFetcher, RssStore},
         session::SessionStore,
+        tools::rss::{RssFetchConfig, RssFetcher, RssStore},
         tools::todo::TodoStore,
         tools::{DynRadarExecutor, build_radar_executor},
         train::{DynTrainExecutor, build_train_executor},
