+++ qq-maid-core/src/runtime/mod.rs
@@ -14,7 +14,6 @@ pub mod pending;
 pub mod prompt;
 pub mod push;
 pub mod respond;
-pub mod rss;
 pub mod session;
 pub mod tools;
 pub mod train;
