+++ qq-maid-core/src/storage/mod.rs
@@ -1,7 +1,7 @@
 //! 数据持久化存储模块。
 //!
-//! 提供长期记忆（memory）、会话（session）、RSS、通知、知识库和 SQLite 数据库能力。
-//! Memory、Session、RSS、Notification、Knowledge 共用项目级 SQLite。SQLite 连接、PRAGMA 和 migration
+//! 提供长期记忆（memory）、会话、通知、知识库和 SQLite 数据库基础能力。
+//! 各业务域共用项目级 SQLite。SQLite 连接、PRAGMA 和 migration
 //! 统一放在 `database` 模块，业务模块只保留自身表结构和查询逻辑。
 
 pub mod database;
@@ -11,7 +11,6 @@ pub mod knowledge;
 pub mod memory;
 pub mod migrations;
 pub mod notification;
-pub mod rss;
 pub mod session;
 
 pub use migrations::APP_MIGRATIONS;
